% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_height.R
\name{convert_height}
\alias{convert_height}
\title{Convert Height}
\usage{
convert_height(cm, age, name)
}
\arguments{
\item{cm}{individual's height in cm}

\item{age}{individual's age (in years)}

\item{name}{individual's name}
}
\description{
A function to convert your height from cm to ft and print along with name and age info
}
\examples{
convert_height(170, 23, "Jordan")
}
\author{
Jordan Sibley
}
